Refactored line 28.
