+++ src/consensus/mod.rs
@@ -119,91 +119,6 @@ impl<A: App<B>, B: Hashable + Clone + Eq> Tendermint<A, B> {
         }
     }
 
-    async fn line28(&mut self) -> Result<bool, Error> {
-        if !self.current.step.is_propose() {
-            return Ok(false);
-        }
-
-        let proposer = self.app.proposer(self.height, self.current.round);
-        let messages = self.log.get_current();
-        let broadcast = messages
-            .proposals
-            .iter()
-            .filter(|contract| contract.signee.hash() == proposer)
-            .map(|contract| &contract.content)
-            .filter(|proposal| {
-                if let Some(vr) = proposal.valid_round {
-                    vr < self.current.round
-                        && proposal.height == self.height
-                        && proposal.round == self.current.round
-                } else {
-                    false
-                }
-            })
-            .find(|proposal| {
-                let id = proposal.proposal.hash();
-                let total_weight = messages
-                    .prevotes
-                    .iter()
-                    .map(|contract| {
-                        (
-                            self.current.voting_weight(contract.signee.hash()),
-                            &contract.content,
-                        )
-                    })
-                    .filter(|(_, prevote)| {
-                        prevote.height == self.height
-                            && prevote.round == proposal.round
-                            && prevote.id == Some(id)
-                    })
-                    .map(|(weight, _)| weight)
-                    .sum::<u64>();
-
-                total_weight > self.current.two_thirds
-            });
-
-        let proposal = match broadcast {
-            Some(contract) => contract,
-            None => return Ok(false),
-        };
-
-        let Proposal {
-            height,
-            round,
-            proposal: ref v,
-            valid_round,
-        } = proposal;
-
-        let valid_round = match valid_round {
-            Some(x) => x,
-            None => return Ok(false),
-        };
-
-        if (height, round) == (&self.height, &self.current.round) {
-            let vote_id = if self.app.validate_block(v)
-                && self
-                    .locked
-                    .as_ref()
-                    .map(|x| &x.round <= valid_round || &x.value == v)
-                    .unwrap_or(true)
-            {
-                Some(v.hash())
-            } else {
-                None
-            };
-
-            let prevote = Prevote::new(self.height, self.current.round, vote_id);
-            self.broadcast(Broadcast::Prevote(self.app.sign(prevote)))
-                .await?;
-
-            self.current.step = Step::prevote();
-
-            Ok(true)
-        } else {
-            Ok(false)
-        }
-    }
-
     async fn line34(&mut self) -> Result<bool, Error> {
         if !matches!(
             self.current.step,
